--- gatsby/src/components/navbar.tsx
@@ -1,150 +1,0 @@
-import * as React from "react"
-import { HomeNavButton, NavButton, CareersNavButton, SvgButton } from "./buttons"
-
-import { isMobile, ResponsiveProps } from "../utils"
-
-import * as styles from "../styles/navbar.module.css"
-import * as orientationStyles from "../styles/utilities/orientation.module.css"
-import { hidden, desktop, mobile } from "../styles/utilities/visibility.module.css"
-
-import Chilidogmenu from "../images/svg/chilidogmenu.svg"
-import Menuclose from "../images/svg/menuclose.svg"
-
-interface NavItemsProps {
-  id: string
-  rootClassName: string
-  navItemClassName: string
-}
-
-function NavItems(props: NavItemsProps) {
-  return (
-    <div className={props.rootClassName}>
-      <NavButton className={props.navItemClassName} href="/services" text="Services" />
-      <NavButton className={props.navItemClassName} href="/work" text="Work" />
-      <NavButton className={props.navItemClassName} href="/articles" text="Articles" />
-      <NavButton className={props.navItemClassName} href="/contact" text="Contact" />
-      <CareersNavButton 
-        className={props.navItemClassName}
-        href="https://careers.superformula.com"
-        text="Careers"
-      />
-    </div>
-  )
-}
-
-function NavMenuDesktop() {
-  const homeBtnDesktop: string = desktop + " " + styles.homeBtn
-  const navDesktop: string = styles.container + " " + styles.navbar + " " + orientationStyles.horizontalFlex
-  const rootDesktop: string = styles.navItem + " " + orientationStyles.horizontalFlex
-  const homeItem: string = styles.homeItem
-  const navItemDesktop: string = styles.navItem
-
-  return (
-    <>
-      <HomeNavButton className={homeBtnDesktop} href="/" text="Superformula." />
-      <div className={navDesktop}>
-        <NavItems
-          id={"desktop-nav"}
-          rootClassName={rootDesktop}
-          homeItemClassName={homeItem}
-          navItemClassName={navItemDesktop}
-        />
-      </div>
-    </>
-  )
-}
-
-interface NavMenuMobileClosedProps {
-  onClick: () => void
-}
-
-function NavMenuMobileClosed(props: NavMenuMobileClosedProps) {
-  const wrapper: string = hidden
-  const chilidogStyle: string = mobile + " " + styles.chilidog
-  const homeBtnMobile: string = mobile + " " + styles.home
-  return (
-    <div className={wrapper}>
-      <SvgButton
-        className={chilidogStyle}
-        onClick={props.onClick}
-        icon={<Chilidogmenu width={32} height={45} />}
-      />
-      <HomeNavButton className={homeBtnMobile} href="/" text="Superformula." />
-    </div>
-  )
-}
-
-interface NavMenuMobileOpenedProps {
-  onClick: () => void
-}
-
-function NavMenuMobileOpened(props: NavMenuMobileOpenedProps) {
-  const rootMobile: string = mobile
-      + " " + styles.navItem
-      + " " + orientationStyles.verticalFlex
-      + " " + styles.rootItemMobile
-  const wrapper: string = styles.navbarContainer
-      + " " + styles.homeItem
-      + " " + orientationStyles.mobileB_desktopH
-  const menuClose: string = styles.menuClose
-  const navItemMobile: string = styles.pageItem + " " + styles.navItem
-  return (
-    <div className={rootMobile} >
-      <SvgButton
-        className={menuClose}
-        onClick={props.onClick}
-        icon={<Menuclose width={40} height={40} />}
-      />
-      <NavItems
-        rootClassName={wrapper}
-        navItemClassName={navItemMobile}
-      />
-    </div>
-  )
-}
-
-function NavMenuMobile() {
-  const [isMobileMenuOpen, setIsMenuOpen] = React.useState(false)
-
-  function onChilidogClick(): void {
-    if (!isMobileMenuOpen) {
-      const menu: HTMLElement | null = document.getElementById("mobile-nav")
-      menu.classList.remove(hidden)
-      setIsMenuOpen(true)
-    }
-  }
-
-  function onMenuCloseClick(): void {
-    if (isMobileMenuOpen) {
-      const menu: HTMLElement | null = document.getElementById("mobile-nav")
-      menu.classList.add(hidden)
-      setIsMenuOpen(false)
-    }
-  }
-
-  const rootMobile: string = mobile
-      + " " + styles.navItem
-      + " " + orientationStyles.verticalFlex
-      + " " + styles.rootItemMobile
-
-  return (
-    <nav id={"mobile-nav"} className={rootMobile}>
-      <NavMenuMobileClosed onClick={onChilidogClick} />
-      <NavMenuMobileOpened onClick={onMenuCloseClick} />
-    </nav>
-  )
-}
-
-export default function NavBar(props: ResponsiveProps) {
-  const navPadding: string = styles.fixed
-      + " " + styles.navbarContainer
-      + " " + orientationStyles.horizontalFlex
-  const navMobile: string = ""
-
-  return (
-    <div id={"desktop-nav-wrapper"} className={navPadding}>
-      <NavMenuDesktop />
-      <NavMenuMobile />
-    </div>
-  )
-}
